Import images Material theme for admin panel

+++ app/Console/Commands/ProductImportCommand.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Commands;
 
+use App\Jobs\JobProductImageDownload;
 use App\Models\Brand;
 use App\Models\Category;
 use App\Models\CategoryTranslate;
@@ -18,7 +19,7 @@ use Illuminate\Support\Str;
 class ProductImportCommand extends Command
 {
 
-    protected $signature = 'product:import {--debug}';
+    protected $signature = 'product:import {--debug} {--full}';
     protected $description = 'Import products from B2B';
 
     public function __construct()
@@ -29,67 +30,71 @@ class ProductImportCommand extends Command
     public function handle()
     {
         $debug = $this->option('debug');
+        $full = $this->option('full');
 
         $start = Carbon::now();
 
-        $this->info('Download json'); echo PHP_EOL;
-
         $json = '{}';
 
-        if($debug) {
+        if ($debug) {
             try {
                 $json = json_decode(Storage::get('public/json.txt'), true);
-            } catch(\Exception $e) {
-                $this->error($e->getMessage()); return 0;
+            } catch (\Exception $e) {
+                $this->error($e->getMessage());
+                return 0;
             }
         } else {
             $response = Http::get('https://b2b-sandi.com.ua/export/view/c2f48b8ccee7e93c4902132271a25236-3982-1652309498/json');
             $json = $response->json();
         }
 
-        // TODO Download images ???
-
         // BRANDS
+        $startSection = Carbon::now();
         $brands = [];
-        $this->newLine(2);
-        $this->info('Brands');
-        $barBrands = $this->output->createProgressBar(count($json['brands'])); echo PHP_EOL;
+        $this->info('Brands: ');
+        $barBrands = $this->output->createProgressBar(count($json['brands']));
+        echo PHP_EOL;
         $barBrands->start();
         DB::beginTransaction();
         try {
             Brand::query()->delete();
-            if(isset($json['brands'])) {
-                foreach($json['brands'] as $ref => $brand) {
+            if (isset($json['brands'])) {
+                foreach ($json['brands'] as $ref => $brand) {
                     $brand['deleted_at'] = null;
                     $b = Brand::withTrashed()->updateOrCreate(['ref' => $ref], $brand);
                     $brands[$b->ref] = $b->id;
                     $barBrands->advance();
                 }
             }
-        } catch(\Exception $e) {
+        } catch (\Exception $e) {
             DB::rollBack();
             $this->error($e->getMessage());
         }
         DB::commit();
+
+        $this->newLine();
         $barBrands->finish();
+        $this->table(['Imported items', 'Time'], [[count($json['brands']), Carbon::now()->diff($startSection)->format('%H:%I:%S')]]);
+        $this->newLine();
 
         // CATEGORIES
+        $startSection = Carbon::now();
         $categories = [];
-        $this->newLine(2);
-        $this->info('Categories');
-        $barsCategories = $this->output->createProgressBar(count($json['categories'])); echo PHP_EOL;
+        $this->info('Categories: ');
+        $barsCategories = $this->output->createProgressBar(count($json['categories']));
+        echo PHP_EOL;
         $barsCategories->start();
         DB::beginTransaction();
         try {
             Category::query()->delete();
-            if(isset($json['categories'])) {
-                foreach($json['categories'] as $ref => $cat) {
+            if (isset($json['categories'])) {
+                foreach ($json['categories'] as $ref => $cat) {
                     $cat['deleted_at'] = null;
                     $category = Category::withTrashed()->updateOrCreate(['ref' => $ref], $cat);
                     $categories[$category->ref] = $category->id;
                     $category->translates()->delete();
                     $pattern = '/([0-9]+)\. /';
-                    foreach($cat['name'] as $lang => $name) {
+                    foreach ($cat['name'] as $lang => $name) {
                         $name = preg_replace($pattern, '', $name);
                         CategoryTranslate::create([
                             'category_id' => $category->id,
@@ -100,18 +105,23 @@ class ProductImportCommand extends Command
                     $barsCategories->advance();
                 }
             }
-        } catch(\Exception $e) {
+        } catch (\Exception $e) {
             DB::rollBack();
             $this->error($e->getMessage());
         }
         DB::commit();
+
+        $this->newLine();
         $barsCategories->finish();
+        $this->table(['Imported items', 'Time'], [[count($json['categories']), Carbon::now()->diff($startSection)->format('%H:%I:%S')]]);
+        $this->newLine();
 
         // ATTRIBUTES
+        $startSection = Carbon::now();
         $attributes = [];
-        $this->newLine(2);
-        $this->info('Attributes');
-        $barAttributes = $this->output->createProgressBar(count($json['attributes'])); echo PHP_EOL;
+        $this->info('Attributes: ');
+        $barAttributes = $this->output->createProgressBar(count($json['attributes']));
+        echo PHP_EOL;
         $barAttributes->start();
         DB::beginTransaction();
         try {
@@ -127,38 +137,48 @@ class ProductImportCommand extends Command
                 $attribute->translates()->createMany($attributeTranslates);
                 $barAttributes->advance();
             }
-        } catch(\Exception $e) {
+        } catch (\Exception $e) {
             DB::rollBack();
             $this->error($e->getMessage());
         }
         DB::commit();
+
+        $this->newLine();
         $barAttributes->finish();
+        $this->table(['Imported items', 'Time'], [[count($json['attributes']), Carbon::now()->diff($startSection)->format('%H:%I:%S')]]);
+        $this->newLine();
 
         // PRODUCTS
-        $this->newLine(2);
-        $this->info('Products');
-        $barsProducts = $this->output->createProgressBar(count($json['products'])); echo PHP_EOL;
+        $startSection = Carbon::now();
+        $this->info('Products: ');
+        $barsProducts = $this->output->createProgressBar(count($json['products']));
+        echo PHP_EOL;
         $barsProducts->start();
-        DB::beginTransaction();
-        try {
-            Product::query()->delete();
-            if (isset($json['products'])) {
-                foreach ($json['products'] as $sku => $p) {
-                    /** CHECK BALANCE */
-                    $balanceCount = collect($p['main']['balances'])->except('213')->sum();
-                    if ($balanceCount < 3) continue;
-                    /** MAIN */
+
+        Product::query()->delete();
+        if (isset($json['products'])) {
+            foreach ($json['products'] as $sku => $p) {
+
+                /** CHECK BALANCE */
+                $balanceCount = collect($p['main']['balances'])->except('213')->sum();
+                if ($balanceCount < 3) continue;
+
+                /** MAIN */
+                if($full) {
                     $p['main']['slug'] = (isset($p['main']['name']['ru'])) ? Str::slug($p['main']['name']['ru'] . '-' . $p['main']['vendorCode']) : (isset($p['main']['name']['uk'])) ? Str::slug($p['main']['name']['uk'] . '-' . $p['main']['vendorCode']) : Str::uuid();
                     $p['main']['brand_id'] = $brands[$p['main']['brand']];
                     $p['main']['brand_ref'] = $p['main']['brand'];
                     $p['main']['category_id'] = $categories[$p['main']['category']];
                     $p['main']['category_ref'] = $p['main']['category'];
-                    $p['main']['balance'] = $balanceCount;
                     $p['main']['deleted_at'] = null;
-                    $product = Product::withTrashed()->updateOrCreate([
-                        'sku' => $sku
-                    ], $p['main']);
-                    /** TRANSLATES */
+                }
+                $p['main']['balance'] = $balanceCount;
+                $product = Product::withTrashed()->updateOrCreate([
+                    'sku' => $sku
+                ], $p['main']);
+
+                /** TRANSLATES */
+                if($full) {
                     $product->translates()->delete();
                     foreach ($p['main']['name'] as $lang => $name) {
                         $translate = [
@@ -168,7 +188,10 @@ class ProductImportCommand extends Command
                         ];
                         $product->translates()->create($translate);
                     }
-                    /** ATTRIBUTES */
+                }
+
+                /** ATTRIBUTES */
+                if($full) {
                     $product->attributes()->delete();
                     foreach ($p['attributes'] as $ref => $attr) {
                         $attribute = ProductToAttribute::create([
@@ -183,40 +206,50 @@ class ProductImportCommand extends Command
                             ]);
                         }
                     }
-                    /** PRICES */
-                    if (isset($p['main']['prices']) and $p['main']['prices']['retail']['current']) {
-                        $price = [
-                            'retail' => $p['main']['prices']['retail']['current'],
-                            'retail_old' => $p['main']['prices']['retail']['old'] ?? null,
-                            'purchase' => $p['main']['prices']['purchase']['cash']['current'] ?? null,
-                            'purchase_old' => $p['main']['prices']['purchase']['cash']['old'] ?? null,
-                        ];
-                        $product->price()->delete();
-                        $product->price()->create($price);
-                    }
-                    /** IMAGES */
-                    $images = [
-                        ['link' => $p['images']['main'], 'order' => 0, 'main' => 1]
-                    ];
-                    foreach ($p['images']['additional'] as $order => $img) {
-                        $images[] = ['link' => $img, 'order' => $order];
-                    }
-                    $product->images()->createMany($images);
-                    /** BALANCES */
-
-
-                    $barsProducts->advance();
                 }
-            }
-        } catch(\Exception $e) {
-            DB::rollBack();
-            $this->error($e->getMessage());
-        }
-        DB::commit();
-        $barsProducts->finish();
 
-        $finish = Carbon::now();
-        $this->newLine(2); $this->info('Done in ' . $start->diff($finish)->format('%i:%s'));
+                /** PRICES */
+                if (isset($p['main']['prices']) and $p['main']['prices']['retail']['current']) {
+                    $price = [
+                        'retail' => $p['main']['prices']['retail']['current'],
+                        'retail_old' => $p['main']['prices']['retail']['old'] ?? null,
+                        'purchase' => $p['main']['prices']['purchase']['cash']['current'] ?? null,
+                        'purchase_old' => $p['main']['prices']['purchase']['cash']['old'] ?? null,
+                    ];
+                    $product->price()->delete();
+                    $product->price()->create($price);
+                }
+
+                /** IMAGES */
+                if($full) {
+                    $product->load('images');
+                    $product->images()->updateOrCreate(['order' => 0], [
+                        'link' => $p['images']['main'], 'order' => 0, 'main' => 1
+                    ]);
+                    foreach ($p['images']['additional'] as $order => $img) {
+                        $product->images()->updateOrCreate(['order' => $order], [
+                            'link' => $img, 'order' => $order, 'main' => 0
+                        ]);
+                    }
+                }
+
+                /** IMAGES DOWNLOADING */
+                if($full) {
+                    JobProductImageDownload::dispatch($product)->onQueue('import_images');
+                }
+
+                $barsProducts->advance();
+            }
+        }
+
+        $this->newLine();
+        $barsProducts->finish();
+        $this->table(['Imported items', 'Time'], [[count($json['products']), Carbon::now()->diff($startSection)->format('%H:%I:%S')]]);
+        $this->newLine();
+
+        $this->info('Import finish');
+        $this->newLine();
+        $this->table(['Time'], [[Carbon::now()->diff($start)->format('%H:%I:%S')]]);
 
         return 0;
     }
